+++ src/main/java/com/lowdragmc/mbd2/MBD2.java
@@ -83,4 +83,8 @@ public class MBD2 {
     public static boolean isKubeJSLoaded() {
         return LDLib.isModLoaded("kubejs");
     }
+
+    public static boolean isAE2Loaded() {
+        return LDLib.isModLoaded("ae2");
+    }
 }