+++ CHANGELOG.md
@@ -4,4 +4,5 @@
 * Fixed Forge Additional Capabilities doesn't work
 * Added Light Condition for `sky light`, `block light` and `can see sky`
 * Added Consume Inputs After Working (furnace-like machine)
+* Added ME Interface Trait